requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.20
pydantic==2.11.0
pandas==2.2.2
faker==37.8.0
faiss-cpu==1.12.0
sentence-transformers==2.2.2
python-multipart==0.0.6
alembic==1.16.5
huggingface-hub==0.16.4
langchain-core==0.3.76
ollama==0.5.4
langchain-community==0.3.29
protobuf==3.20.3
protobuf==4.21.12
sqlalchemy==2.0.29